remove adapter for v0 and v1
