+++ .github/workflows/continuous-integration.yml
@@ -57,14 +57,3 @@ jobs:
         with:
           name: Test report
           path: build/reports/tests/test
-  build-with-nullaway-serialization-0:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up JDK 11
-        uses: actions/setup-java@v2
-        with:
-          java-version: 11
-          distribution: 'temurin'
-      - name: Build with Gradle
-        run: ./gradlew :annotator-core:test --scan -Pnullaway-serialization-format-version=0
