+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/adapters/NullAwayV2Adapter.java
@@ -5,14 +5,36 @@ import edu.ucr.cs.riple.core.Config;
 import edu.ucr.cs.riple.core.metadata.field.FieldDeclarationStore;
 import edu.ucr.cs.riple.core.metadata.index.Error;
 import edu.ucr.cs.riple.core.metadata.trackers.Region;
+import edu.ucr.cs.riple.core.metadata.trackers.TrackerNode;
 import edu.ucr.cs.riple.injector.location.Location;
+import edu.ucr.cs.riple.injector.location.OnField;
+import edu.ucr.cs.riple.scanner.generatedcode.SourceType;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Set;
+import java.util.stream.Collectors;
 
-public class NullAwayV2Adapter extends NullAwayV1Adapter {
+public class NullAwayV2Adapter extends NullAwayAdapterBaseClass {
 
   public NullAwayV2Adapter(Config config, FieldDeclarationStore fieldDeclarationStore) {
     super(config, fieldDeclarationStore);
+  }
+
+  @Override
+  public TrackerNode deserializeTrackerNode(String[] values) {
+    Preconditions.checkArgument(
+        values.length == 5,
+        "Expected 5 values to create TrackerNode instance in NullAway serialization version 1 but found: "
+            + values.length);
+    return new TrackerNode(
+        new Region(values[0], values[1], SourceType.valueOf(values[4])), values[2], values[3]);
+  }
+
+  @Override
+  public Set<Region> getFieldRegionScope(OnField onField) {
+    return onField.variables.stream()
+        .map(fieldName -> new Region(onField.clazz, fieldName))
+        .collect(Collectors.toSet());
   }
 
   @Override
